refactor(ast): Move back to original module path
